update: return API embedding response
Originally Embeddings method would return an object
which was a result of a conversion to a smaller type which only
contained the embeddings. This prevents the clients from accessing the
API response fields which is not ideal.

This commit reverts to returning deserialized API responses and instead
provides an exported func in each package that lets the consumers to
convert the object into a higher level embedding type defined in the
module root.

Signed-off-by: Milos Gajdos <milosthegajdos@gmail.com>